Feat(todo):todo 기능 추가

+++ src/repositories/todolist.respository.ts
@@ -30,7 +30,7 @@ export class ToDoListRepository {
             .getOne();
     }
 
-    async findAllByDateAndUserId(userId: number, date: string): Promise<ToDoList[]> {
+    async getAllByLikeDate(userId: number, date: string): Promise<ToDoList[]> {
         return await this.todoListRepository
             .createQueryBuilder('tl')
             .innerJoinAndSelect('tl.user', 'u')
@@ -40,11 +40,23 @@ export class ToDoListRepository {
             .getMany();
     }
 
-    async delete(todoId: number): Promise<void> {
+    async getAllByDate(userId: number, date: string): Promise<ToDoList[]> {
+        return await this.todoListRepository
+            .createQueryBuilder('tl')
+            .innerJoinAndSelect('tl.user', 'u')
+            .innerJoinAndSelect('tl.calendar', 'ca')
+            .where('tl.date = :date', { date: date })
+            .andWhere('u.id = :userId', { userId: userId })
+            .getMany();
+    }
+
+    async delete(userId: number, todoId: number): Promise<void> {
         await this.todoListRepository
             .createQueryBuilder('tl')
             .delete()
             .where('id = :todoId', { todoId: todoId })
+            .andWhere('userId = :userId', { userId: userId })
             .execute();
+        return;
     }
 }
